backend/requirements.txt
Django==6.0.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
psycopg[binary]==3.3.3
gunicorn==23.0.0
python-dotenv==1.0.1
dj-database-url==2.3.0
whitenoise==6.8.2
Pillow==11.1.0
